requirements.txt
datasets
streamlit
langchain-core
langchain
langchain-community
python-dotenv
langchain_openai
neo4j
numpy
ragas
retry
rank_bm25
lark
pymongo
langchainhub
langchain-pinecone
langchain_postgres
langchain_elasticsearch
motor
fastapi
pymupdf
python-docx
paddleocr
paddlepaddle
openpyxl
pytz
xlrd
pinecone
redis
cohere
langchain-cohere
langchain-cli
scikit-learn
PyPDF2